move all tests of train/eval mode to test_train_eval_mode
