+++ tests/test_run.py
@@ -1,13 +1,5 @@
 import numpy as np
 import pytest
-import torch as th
-from tests.test_predict import (
-    FlattenBatchNormDropoutExtractor,
-    clone_batch_norm_stats,
-    clone_dqn_batch_norm_stats,
-    clone_sac_batch_norm_stats,
-    clone_td3_batch_norm_stats,
-)
 
 from stable_baselines3 import A2C, DDPG, DQN, PPO, SAC, TD3
 from stable_baselines3.common.noise import NormalActionNoise, OrnsteinUhlenbeckActionNoise
@@ -147,312 +139,3 @@ def test_train_freq_fail(train_freq):
             train_freq=train_freq,
         )
         model.learn(total_timesteps=250)
-
-
-def test_dqn_train_with_batch_norm():
-    model = DQN(
-        "MlpPolicy",
-        "CartPole-v1",
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        learning_starts=0,
-        tau=0,
-        seed=1,
-    )
-
-    (
-        q_net_bias_before,
-        q_net_running_mean_before,
-        q_net_target_bias_before,
-        q_net_target_running_mean_before,
-    ) = clone_dqn_batch_norm_stats(model)
-
-    model.learn(total_timesteps=200)
-
-    (
-        q_net_bias_after,
-        q_net_running_mean_after,
-        q_net_target_bias_after,
-        q_net_target_running_mean_after,
-    ) = clone_dqn_batch_norm_stats(model)
-
-    assert ~th.isclose(q_net_bias_before, q_net_bias_after).all()
-    assert ~th.isclose(q_net_running_mean_before, q_net_running_mean_after).all()
-
-    assert th.isclose(q_net_target_bias_before, q_net_target_bias_after).all()
-    assert th.isclose(q_net_target_running_mean_before, q_net_target_running_mean_after).all()
-
-
-def test_dqn_collect_rollouts_with_batch_norm():
-    model = DQN(
-        "MlpPolicy",
-        "CartPole-v1",
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        seed=1,
-    )
-
-    (
-        q_net_bias_before,
-        q_net_running_mean_before,
-        q_net_target_bias_before,
-        q_net_target_running_mean_before,
-    ) = clone_dqn_batch_norm_stats(model)
-
-    total_timesteps, callback = model._setup_learn(total_timesteps=100, eval_env=model.get_env())
-
-    for _ in range(10):
-        model.collect_rollouts(
-            model.get_env(),
-            train_freq=model.train_freq,
-            action_noise=model.action_noise,
-            callback=callback,
-            learning_starts=model.learning_starts,
-            replay_buffer=model.replay_buffer,
-        )
-
-    (
-        q_net_bias_after,
-        q_net_running_mean_after,
-        q_net_target_bias_after,
-        q_net_target_running_mean_after,
-    ) = clone_dqn_batch_norm_stats(model)
-
-    assert th.isclose(q_net_bias_before, q_net_bias_after).all()
-    assert th.isclose(q_net_running_mean_before, q_net_running_mean_after).all()
-
-    assert th.isclose(q_net_target_bias_before, q_net_target_bias_after).all()
-    assert th.isclose(q_net_target_running_mean_before, q_net_target_running_mean_after).all()
-
-
-def test_td3_train_with_batch_norm():
-    model = TD3(
-        "MlpPolicy",
-        "Pendulum-v0",
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        learning_starts=0,
-        tau=0,
-        seed=1,
-    )
-
-    (
-        actor_bias_before,
-        actor_running_mean_before,
-        critic_bias_before,
-        critic_running_mean_before,
-        actor_target_bias_before,
-        actor_target_running_mean_before,
-        critic_target_bias_before,
-        critic_target_running_mean_before,
-    ) = clone_td3_batch_norm_stats(model)
-
-    model.learn(total_timesteps=200)
-
-    (
-        actor_bias_after,
-        actor_running_mean_after,
-        critic_bias_after,
-        critic_running_mean_after,
-        actor_target_bias_after,
-        actor_target_running_mean_after,
-        critic_target_bias_after,
-        critic_target_running_mean_after,
-    ) = clone_td3_batch_norm_stats(model)
-
-    assert ~th.isclose(actor_bias_before, actor_bias_after).all()
-    assert ~th.isclose(actor_running_mean_before, actor_running_mean_after).all()
-
-    assert ~th.isclose(critic_bias_before, critic_bias_after).all()
-    assert ~th.isclose(critic_running_mean_before, critic_running_mean_after).all()
-
-    assert th.isclose(actor_target_bias_before, actor_target_bias_after).all()
-    assert th.isclose(actor_target_running_mean_before, actor_target_running_mean_after).all()
-
-    assert th.isclose(critic_target_bias_before, critic_target_bias_after).all()
-    assert th.isclose(critic_target_running_mean_before, critic_target_running_mean_after).all()
-
-
-def test_td3_collect_rollouts_with_batch_norm():
-    model = TD3(
-        "MlpPolicy",
-        "Pendulum-v0",
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        seed=1,
-    )
-
-    (
-        actor_bias_before,
-        actor_running_mean_before,
-        critic_bias_before,
-        critic_running_mean_before,
-        actor_target_bias_before,
-        actor_target_running_mean_before,
-        critic_target_bias_before,
-        critic_target_running_mean_before,
-    ) = clone_td3_batch_norm_stats(model)
-
-    total_timesteps, callback = model._setup_learn(total_timesteps=100, eval_env=model.get_env())
-
-    for _ in range(10):
-        model.collect_rollouts(
-            model.get_env(),
-            train_freq=model.train_freq,
-            action_noise=model.action_noise,
-            callback=callback,
-            learning_starts=model.learning_starts,
-            replay_buffer=model.replay_buffer,
-        )
-
-    (
-        actor_bias_after,
-        actor_running_mean_after,
-        critic_bias_after,
-        critic_running_mean_after,
-        actor_target_bias_after,
-        actor_target_running_mean_after,
-        critic_target_bias_after,
-        critic_target_running_mean_after,
-    ) = clone_td3_batch_norm_stats(model)
-
-    assert th.isclose(actor_bias_before, actor_bias_after).all()
-    assert th.isclose(actor_running_mean_before, actor_running_mean_after).all()
-
-    assert th.isclose(critic_bias_before, critic_bias_after).all()
-    assert th.isclose(critic_running_mean_before, critic_running_mean_after).all()
-
-    assert th.isclose(actor_target_bias_before, actor_target_bias_after).all()
-    assert th.isclose(actor_target_running_mean_before, actor_target_running_mean_after).all()
-
-    assert th.isclose(critic_target_bias_before, critic_target_bias_after).all()
-    assert th.isclose(critic_target_running_mean_before, critic_target_running_mean_after).all()
-
-
-def test_sac_train_with_batch_norm():
-    model = SAC(
-        "MlpPolicy",
-        "Pendulum-v0",
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        learning_starts=0,
-        tau=0,
-        seed=1,
-    )
-
-    (
-        actor_bias_before,
-        actor_running_mean_before,
-        critic_bias_before,
-        critic_running_mean_before,
-        critic_target_bias_before,
-        critic_target_running_mean_before,
-    ) = clone_sac_batch_norm_stats(model)
-
-    model.learn(total_timesteps=200)
-
-    (
-        actor_bias_after,
-        actor_running_mean_after,
-        critic_bias_after,
-        critic_running_mean_after,
-        critic_target_bias_after,
-        critic_target_running_mean_after,
-    ) = clone_sac_batch_norm_stats(model)
-
-    assert ~th.isclose(actor_bias_before, actor_bias_after).all()
-    assert ~th.isclose(actor_running_mean_before, actor_running_mean_after).all()
-
-    assert ~th.isclose(critic_bias_before, critic_bias_after).all()
-    assert ~th.isclose(critic_running_mean_before, critic_running_mean_after).all()
-
-    assert th.isclose(critic_target_bias_before, critic_target_bias_after).all()
-    assert th.isclose(critic_target_running_mean_before, critic_target_running_mean_after).all()
-
-
-def test_sac_collect_rollouts_with_batch_norm():
-    model = SAC(
-        "MlpPolicy",
-        "Pendulum-v0",
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        seed=1,
-    )
-
-    (
-        actor_bias_before,
-        actor_running_mean_before,
-        critic_bias_before,
-        critic_running_mean_before,
-        critic_target_bias_before,
-        critic_target_running_mean_before,
-    ) = clone_sac_batch_norm_stats(model)
-
-    total_timesteps, callback = model._setup_learn(total_timesteps=100, eval_env=model.get_env())
-
-    for _ in range(10):
-        model.collect_rollouts(
-            model.get_env(),
-            train_freq=model.train_freq,
-            action_noise=model.action_noise,
-            callback=callback,
-            learning_starts=model.learning_starts,
-            replay_buffer=model.replay_buffer,
-        )
-
-    (
-        actor_bias_after,
-        actor_running_mean_after,
-        critic_bias_after,
-        critic_running_mean_after,
-        critic_target_bias_after,
-        critic_target_running_mean_after,
-    ) = clone_sac_batch_norm_stats(model)
-
-    assert th.isclose(actor_bias_before, actor_bias_after).all()
-    assert th.isclose(actor_running_mean_before, actor_running_mean_after).all()
-
-    assert th.isclose(critic_bias_before, critic_bias_after).all()
-    assert th.isclose(critic_running_mean_before, critic_running_mean_after).all()
-
-    assert th.isclose(critic_target_bias_before, critic_target_bias_after).all()
-    assert th.isclose(critic_target_running_mean_before, critic_target_running_mean_after).all()
-
-
-@pytest.mark.parametrize("model_class", [A2C, PPO])
-@pytest.mark.parametrize("env_id", ["Pendulum-v0", "CartPole-v1"])
-def test_a2c_ppo_train_with_batch_norm(model_class, env_id):
-    model = model_class(
-        "MlpPolicy",
-        env_id,
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        seed=1,
-    )
-
-    batch_norm = model.policy.features_extractor.batch_norm
-    bias_before, running_mean_before = clone_batch_norm_stats(batch_norm)
-
-    model.learn(total_timesteps=200)
-
-    bias_after, running_mean_after = clone_batch_norm_stats(batch_norm)
-
-    assert ~th.isclose(bias_before, bias_after).all()
-    assert ~th.isclose(running_mean_before, running_mean_after).all()
-
-
-@pytest.mark.parametrize("model_class", [A2C, PPO])
-@pytest.mark.parametrize("env_id", ["Pendulum-v0", "CartPole-v1"])
-def test_a2c_ppo_collect_rollouts_with_batch_norm(model_class, env_id):
-    model = model_class(
-        "MlpPolicy",
-        env_id,
-        policy_kwargs=dict(net_arch=[16, 16], features_extractor_class=FlattenBatchNormDropoutExtractor),
-        seed=1,
-    )
-
-    batch_norm = model.policy.features_extractor.batch_norm
-    bias_before, running_mean_before = clone_batch_norm_stats(batch_norm)
-
-    total_timesteps, callback = model._setup_learn(total_timesteps=100, eval_env=model.get_env())
-
-    for _ in range(10):
-        model.collect_rollouts(model.get_env(), callback, model.rollout_buffer, n_rollout_steps=model.n_steps)
-
-    bias_after, running_mean_after = clone_batch_norm_stats(batch_norm)
-
-    assert th.isclose(bias_before, bias_after).all()
-    assert th.isclose(running_mean_before, running_mean_after).all()
